+++ bmx280/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.4] - 2017-08-03
+### Added
+- Humidity support for the BME280 class of sensors
+### Updated
+- added required=false to uses-library on manifests
+
 ## [0.3] - 2017-06-16
 ### Fixed
 - optional I2C address to driver constructor (issue #40)